requirements.txt
google-play-scraper==1.2.7
langdetect==1.0.9
emoji==2.12.1
python-dotenv==1.0.1
groq==0.13.1
google-genai==1.7.0
pytest==8.3.5
